+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Fixed
 
+* Fix a crash in graph filtering module when not removing logging handlers.
 * Bux fix in graph filtering module: filtering cell area for at least one cell was filtering on all cells instead of at least one cell.
 * Small bugs in Registration module